create separate module for publisher
