+++ bitmex-publisher/src/main/java/com/mycompany/app/Publisher.java
@@ -1,9 +1,10 @@
-package com.BitmexStreamPublisher;
+package com.mycompany.app;
 
 import com.google.api.core.ApiFuture;
 import com.google.protobuf.ByteString;
 import com.google.pubsub.v1.PubsubMessage;
 import com.google.pubsub.v1.TopicName;
+
 import java.io.IOException;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
