--- src/main/java/com/BitmexStreamSubscriber/BitmexSubscriber.java
@@ -1,4 +1,0 @@
-package com.BitmexStreamSubscriber;
-
-public class BitmexSubscriber {
-}
